moved input_info utility methods to separate module
