+++ sciencebeam_trainer_delft/sequence_labelling/grobid_trainer.py
@@ -3,9 +3,9 @@ import logging
 import argparse
 import time
 from abc import abstractmethod
-from collections import Counter, defaultdict, OrderedDict
+from collections import Counter
 from itertools import islice
-from typing import Dict, List, Tuple
+from typing import List, Tuple
 
 import sciencebeam_trainer_delft.utils.no_warn_if_disabled  # noqa, pylint: disable=unused-import
 import sciencebeam_trainer_delft.utils.no_keras_backend_message  # noqa, pylint: disable=unused-import
@@ -34,6 +34,14 @@ from sciencebeam_trainer_delft.sequence_labelling.tag_formatter import (
     TAG_OUTPUT_FORMATS,
     get_tag_result,
     format_tag_result
+)
+
+from sciencebeam_trainer_delft.sequence_labelling.input_info import (
+    iter_flat_features,
+    get_feature_counts,
+    get_suggested_feature_indices,
+    format_dict,
+    format_indices
 )
 
 from sciencebeam_trainer_delft.utils.cli import (
@@ -456,75 +464,6 @@ def tag_input(
     )
     LOGGER.info('tag_result:')
     print(formatted_tag_result)
-
-
-def iter_flat_features(features: np.array):
-    return (
-        features_vector
-        for features_doc in features
-        for features_vector in features_doc
-    )
-
-
-def get_feature_value_counts_by_index(features: np.array, max_feature_values: int = 1000):
-    feature_value_counts_by_index = defaultdict(Counter)
-    for feature_vector in iter_flat_features(features):
-        for index, value in enumerate(feature_vector):
-            feature_value_counts = feature_value_counts_by_index[index]
-            if (
-                    len(feature_value_counts) >= max_feature_values
-                    and value not in feature_value_counts):
-                continue
-            feature_value_counts[value] += 1
-    return feature_value_counts_by_index
-
-
-def get_feature_counts(features: np.array):
-    feature_value_counts_by_index = get_feature_value_counts_by_index(features)
-    return OrderedDict([
-        (index, len(feature_value_counts_by_index[index]))
-        for index in sorted(feature_value_counts_by_index.keys())
-    ])
-
-
-def get_suggested_feature_indices(feature_counts: Dict[int, int], threshold: int = 12):
-    return [
-        index
-        for index in sorted(feature_counts.keys())
-        if feature_counts[index] <= threshold
-    ]
-
-
-def format_dict(d: dict) -> str:
-    return ', '.join([
-        '%s: %s' % (key, value)
-        for key, value in d.items()
-    ])
-
-
-def iter_index_groups(indices: List[int]) -> str:
-    group = []
-    for index in indices:
-        if not group or group[-1] + 1 == index:
-            group.append(index)
-            continue
-        yield group
-        group = []
-    if group:
-        yield group
-
-
-def iter_formatted_index_groups(indices: List[int]) -> str:
-    group = []
-    for group in iter_index_groups(indices):
-        if len(group) == 1:
-            yield str(group[0])
-            continue
-        yield '%s-%s' % (group[0], group[-1])
-
-
-def format_indices(indices: List[int]) -> str:
-    return ','.join(list(iter_formatted_index_groups(indices)))
 
 
 def print_input_info(
